requirements.txt
discord
pandas
python-dotenv
requests